+++ packages/components/src/informative/index.js
@@ -17,4 +17,5 @@ export * from './UserCards';
 export * from './UserDisplayItemList';
 export * from './ChatMessage';
 export * from './ProgressColorBar';
+export * from './ProgressBottomBar';
 export * from './ProgressRing';