+++ packages/leemons/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @bubbles-ui/leemons
 
+## 1.2.122
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.122
+  - @bubbles-ui/calendars@1.2.122
+  - @bubbles-ui/editors@1.2.122
+  - @bubbles-ui/icons@1.2.122
+
+## 1.2.121
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.121
+  - @bubbles-ui/calendars@1.2.121
+  - @bubbles-ui/editors@1.2.121
+  - @bubbles-ui/icons@1.2.121
+
+## 1.2.120
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.120
+  - @bubbles-ui/calendars@1.2.120
+  - @bubbles-ui/editors@1.2.120
+  - @bubbles-ui/icons@1.2.120
+
 ## 1.2.119
 
 ### Patch Changes